partials consolidated
