+++ app/scripts/components/home/Tech.js
